backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-jose[cryptography]
pydantic
prometheus-api-client # optional: helpful Prometheus client
python-dotenv
prometheus-fastapi-instrumentator
sqlalchemy
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
slowapi
psutil
docker==7.1.0
kubernetes
psycopg2-binary
httpx
